oem/requirements.txt
Django==1.9.3
django-flat-theme==1.1.3
django-material==0.6.0
django-nested-admin==2.2.1
six==1.10.0
wheel==0.26.0
